#395 Rename: コンポーネントの名前を変更
